Start with initialize logging
