+++ crates/pixi-build-rattler-build/src/protocol.rs
@@ -195,8 +195,6 @@ impl Protocol for RattlerBuildBackend {
     }
 
     async fn conda_build(&self, params: CondaBuildParams) -> miette::Result<CondaBuildResult> {
-        log_conda_build(&self.config, &params).await?;
-
         // Create the work directory if it does not exist
         tokio_fs::create_dir_all(&params.work_directory)
             .await
@@ -334,7 +332,6 @@ impl ProtocolInstantiator for RattlerBuildBackendInstantiator {
         } else {
             RattlerBuildBackendConfig::default()
         };
-        log_initialize(&config, params.project_model).await?;
 
         let instance = RattlerBuildBackend::new(
             params.manifest_path.as_path(),
@@ -363,31 +360,6 @@ pub(crate) fn default_capabilities() -> BackendCapabilities {
             pixi_build_types::VersionedProjectModel::highest_version(),
         ),
     }
-}
-
-async fn log_initialize(
-    config: &RattlerBuildBackendConfig,
-    project_model: Option<VersionedProjectModel>,
-) -> miette::Result<()> {
-    let Some(ref debug_dir) = config.debug_dir else {
-        return Ok(());
-    };
-
-    let project_model = project_model
-        .ok_or_else(|| miette::miette!("project model is required if debug_dir is given"))?
-        .into_v1()
-        .ok_or_else(|| miette::miette!("project model needs to be v1"))?;
-
-    let project_model_json = serde_json::to_string_pretty(&project_model)
-        .into_diagnostic()
-        .context("failed to serialize project model to JSON")?;
-
-    let project_model_path = debug_dir.join("project_model.json");
-    tokio_fs::write(&project_model_path, project_model_json)
-        .await
-        .into_diagnostic()
-        .context("failed to write project model JSON to file")?;
-    Ok(())
 }
 
 #[cfg(test)]
